requirements.txt
aiohttp==3.10.5
autocorrect==2.6.1
av==13.0.0
colorama==0.4.6
easyocr==1.7.2
hydra-core==1.3.2
loguru==0.7.2
numpy==2.2.3
omegaconf==2.3.0
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==2.2.3
Pillow==11.1.0
pytesseract==0.3.13
python-dotenv==1.0.1
python_Levenshtein==0.26.1
scikit_learn==1.6.1
skimage==0.0
torch==2.6.0
torchmetrics==1.4.2
tqdm==4.65.2
transformers==4.49.0
wandb==0.19.7
